Add Tests, change to class
Changed the inventory.py classes into a class. This was not really necessary for the app but it did make it a lot easier to make the tests.  Gave easy access to the spreadsheet to make a test worksheet to work with.

For the tests, didn't quite end up proper unit testing.  The functions are too dependent since they are working on maintaining an outside source. But what I did was still useful for knowing I'm not breaking functions and for starting to explore more edge cases.